refactor(renderer): extract useInputHistory hook (BET-63)
Pulls terminal-style prompt-history navigation (promptHistory derivation,
navigateHistory Up/Down cycling, updateInputWithHistoryReset) out of the
ChatPanel container into hooks/useInputHistory.ts with an explicit injected-
dep surface (messages, inputRef, setInput, setTypeahead, updateInput). It
never touches the SSE/pin/drain core. Adds a render-harness test: ArrowUp on
an empty composer recalls the last user prompt.
